Moved readFile method to its own class (FileReader.java)
