+++ OSIsoft_Demo/src/OSIsoftServer.java
@@ -1,10 +1,4 @@
 import java.io.File;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.OutputStream;
-
-import javax.microedition.io.Connector;
-import javax.microedition.io.file.FileConnection;
 
 import com.ewon.ewonitf.EWException;
 import com.ewon.ewonitf.ScheduledActionManager;
@@ -63,7 +57,7 @@ public class OSIsoftServer {
       }
       
       //Parse the JSON response and retrieve the JSON Array of items
-      JSONTokener JsonT = new JSONTokener(readFile("file://" +responseFilename));
+      JSONTokener JsonT = new JSONTokener(FileReader.readFile("file://" +responseFilename));
       JSONObject requestResponse = new JSONObject(JsonT);
       JSONArray items = requestResponse.getJSONArray("Items");
       
@@ -91,7 +85,7 @@ public class OSIsoftServer {
          }
          
          //Parse the JSON response and retrieve the JSON Array of items
-         JsonT = new JSONTokener(readFile("file://" +responseFilename));
+         JsonT = new JSONTokener(FileReader.readFile("file://" +responseFilename));
          requestResponse = new JSONObject(JsonT);
          items = requestResponse.getJSONArray("Items");
          if (items.length() > 0) {
@@ -141,32 +135,5 @@ public class OSIsoftServer {
             "  \"Future\": false\r\n" + 
             "}";
       return jsonBody;
-   }
-   
-   // Reads a file and returns it as a string
-   public static String readFile(String path) {
-
-      FileConnection fconn = null;
-      InputStream is       = null;
-      int bufferSize       = 10000;
-
-      try {
-         fconn = (FileConnection) Connector.open(path);
-         // If no exception is thrown, then the URI is valid, but the file may or may not
-         // exist.
-
-         is = fconn.openInputStream();
-         int Read;
-         byte[] Buffer = new byte[bufferSize];
-
-         Read = is.read(Buffer);
-         fconn.close();
-         return new String(Buffer, 0, Read);
-
-      } catch (IOException ioe) {
-         System.out.println("Error: " + ioe.toString());
-         //Returns an empty string in error case
-         return "";
-      }
    }
 }
